Refactor `Interactions` namespace structure: move button-related types to `Buttons` subnamespace, introduce `ComponentInteractionException` base class, and update button ID exceptions.
