+++ Logic/Interactions/Buttons/Exceptions/ButtonIdFormatException.cs
@@ -1,0 +1,10 @@
+﻿namespace Logic.Interactions.Buttons.Exceptions;
+
+public class ButtonIdFormatException : ComponentInteractionException
+{
+    public ButtonIdFormatException(string buttonId, string message)
+    : base("", message)
+    {
+        throw new NotImplementedException();
+    }
+}
