+++ Logic/Interactions/Buttons/Exceptions/ComponentInteractionException.cs
@@ -1,0 +1,12 @@
+﻿namespace Logic.Interactions.Buttons.Exceptions;
+
+public abstract class ComponentInteractionException : Exception
+{
+    string ResourceKey { get; }
+
+    protected ComponentInteractionException(string resourceKey, string message)
+        : base(message)
+    {
+        ResourceKey = resourceKey;
+    }
+}
