+++ Logic/Interactions/Buttons/Exceptions/PrefixFormatException.cs
@@ -1,0 +1,12 @@
+﻿namespace Logic.Interactions.Buttons.Exceptions;
+
+public class PrefixFormatException : ComponentInteractionException
+{
+    public string Prefix { get; private set; }
+
+    public PrefixFormatException(string prefix, string message)
+    : base("", message)
+    {
+        Prefix = prefix;
+    }
+}
