--- Logic/Interactions/Exceptions/ButtonIdFormatException.cs
@@ -1,9 +1,0 @@
-﻿namespace Logic.Interactions.Exceptions;
-
-public class ButtonIdFormatException : Exception
-{
-    public ButtonIdFormatException(string buttonId)
-    {
-        throw new NotImplementedException();
-    }
-}
