+++ Logic/Logic.csproj
@@ -11,7 +11,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <Folder Include="Interactions\Handlers\" />
+      <Folder Include="Interactions\Buttons\Handlers\" />
     </ItemGroup>
 
 </Project>
